git-svn-id: svn://192.168.1.10/public/CoffeeMud@11102 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Items/interfaces/Wearable.java
@@ -776,7 +776,7 @@ public interface Wearable extends Environmental
 			if(allCodes.length>61) return;
 			long newCode = 0;
 			if(allCodes.length>0)
-				newCode = 1<<(allCodes.length-1);
+				newCode = (long)1<<(allCodes.length-1);
 			allCodes=Arrays.copyOf(allCodes, allCodes.length+1);
 			allCodes[allCodes.length-1]=newCode;
 			descs=Arrays.copyOf(descs, descs.length+1);
